+++ x/twap/logic.go
@@ -106,18 +106,16 @@ func (k Keeper) EndBlock(ctx sdk.Context) {
 // with updated spot prices and spot price errors, if any.
 // Returns nil on success.
 // Returns error if:
-// - fails to get previous records.
-// - fails to get denoms from the pool.
-// - the number of records does not match expected relative to the
-//  number of denoms in the pool.
+//   - fails to get previous records.
+//   - fails to get denoms from the pool.
+//   - the number of records does not match expected relative to the
+//     number of denoms in the pool.
 func (k Keeper) updateRecords(ctx sdk.Context, poolId uint64) error {
 	// Will only err if pool doesn't have most recent entry set
 	records, err := k.getAllMostRecentRecordsForPool(ctx, poolId)
 	if err != nil {
 		return err
 	}
-	// TODO: Add a safety assert, that # of records is as we expect, given # of denoms in the pool
-	// namely, that for `k` denoms in pool, there should be k * (k - 1) / 2 records
 
 	denoms, err := k.ammkeeper.GetPoolDenoms(ctx, poolId)
 	if err != nil {